+++ Public/Get-LMSession.ps1
@@ -1,5 +1,6 @@
 Function Get-LMSession {
     Param (
+        [Parameter(Mandatory)]
         [String]$AccountName
     )