+++ Public/Invoke-DeviceDedupe.ps1
@@ -1,3 +1,25 @@
+<#
+.SYNOPSIS
+List and/or remove duplicte devices from a portal based on a specified device group and set of exclusion criteria.
+
+.DESCRIPTION
+List and/or remove duplicte devices from a portal based on a specified device group and set of exclusion criteria.
+
+.EXAMPLE
+Invoke-DeviceDedupe -ListDuplicates -DeviceGroupId 8
+
+.NOTES
+Additional arrays can be specified to exclude certain IPs, sysname and devicetypes from being used for duplicate comparison
+
+.INPUTS
+None. Does not accept pipeline input.
+
+.LINK
+Module repo: https://github.com/stevevillardi/Logic.Monitor.SE
+
+.LINK
+PSGallery: https://www.powershellgallery.com/packages/Logic.Monitor.SE
+#>
 Function Invoke-DeviceDedupe {
 
     [CmdletBinding()]